Refactoring, minor style tweaks
