+++ src/apps/project-collaboration/DeleteMember/index.ts
@@ -1,0 +1,1 @@
+export * from './DeleteMember';
