+++ src/apps/project-collaboration/GroupSelector/index.ts
@@ -1,0 +1,1 @@
+export * from './GroupSelector';
